Implement YouTube OAuth continuation in chat flow with detailed logging
